[NET_HANDLER] Refactored tcp_conn.c into two files
- Renamed tcp_conn.* to connection.* (no UDP and TCP any more)
- Moved all helper functions for sending and receiving messages to message.*
- Pulled logic out of recv_new_msg into separate functions
- Made necessary changes in includes to compile net_handler
