+++ net_handler/README.md
@@ -7,7 +7,8 @@ The network handler's job is to communicate with Dawn and Shepherd through socke
 * `pbc_gen/` - the corresponding C code that is generated from the Protobufs using protobuf-c
 * `net_handler.c` - main entry file that accepts TCP connections from Dawn/Shepherd
 * `net_util.c` - helper functions to communicate with the TCP sockets
-* `tcp_conn.c` - handles the TCP connection with a specific client
+* `connection.c` - handles the connection over TCP with a specific client
+* `message.c` - handles the processing of incoming messages and constructing messages to send to/from a client
 
 ## Building
